Add and edit online presences.
