+++ app/src/main/java/org/permanent/permanent/ui/public/EditMilestoneFragment.kt
@@ -1,36 +1,28 @@
-package org.permanent.permanent.ui.publicWorkspace
+package org.permanent.permanent.ui.public
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.navigation.fragment.findNavController
 import kotlinx.android.synthetic.main.activity_main.*
 import org.permanent.permanent.R
-import org.permanent.permanent.databinding.FragmentMilestoneListBinding
+import org.permanent.permanent.databinding.FragmentEditMilestoneBinding
 import org.permanent.permanent.ui.PermanentBaseFragment
 import org.permanent.permanent.viewmodels.PublicProfileViewModel
 
-class MilestoneListFragment: PermanentBaseFragment(){
+class PublicProfileEditMilestonesFragment: PermanentBaseFragment(){
     private lateinit var viewModel: PublicProfileViewModel
-    private lateinit var binding: FragmentMilestoneListBinding
+    private lateinit var binding: FragmentEditMilestoneBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentMilestoneListBinding.inflate(inflater, container, false)
+        binding = FragmentEditMilestoneBinding.inflate(inflater, container, false)
         binding.executePendingBindings()
         binding.lifecycleOwner = this
-        activity?.toolbar?.menu?.findItem(R.id.plusItem)?.isVisible = true
-
-        activity?.toolbar?.setOnMenuItemClickListener {
-            if (it.itemId == R.id.plusItem) {
-                requireParentFragment().findNavController().navigate(R.id.action_milestonesListFragment_to_editMilestonesFragment)
-            }
-            true
-        }
+        activity?.toolbar?.menu?.findItem(R.id.settingsItem)?.isVisible = false
 
         return binding.root
     }
@@ -40,5 +32,5 @@ class MilestoneListFragment: PermanentBaseFragment(){
 
     override fun disconnectViewModelEvents() {
     }
-}
 
+}
