+++ app/src/main/java/org/permanent/permanent/ui/public/LocationSearchFragment.kt
@@ -1,4 +1,4 @@
-package org.permanent.permanent.ui.publicWorkspace
+package org.permanent.permanent.ui.public
 
 import android.os.Bundle
 import android.view.*
