+++ app/src/main/java/org/permanent/permanent/ui/public/OnlinePresenceListener.kt
@@ -1,4 +1,4 @@
-package org.permanent.permanent.ui.publicWorkspace
+package org.permanent.permanent.ui.public
 
 import org.permanent.permanent.models.ProfileItem
