+++ app/src/main/java/org/permanent/permanent/ui/public/PublicGalleryFragment.kt
@@ -1,28 +1,25 @@
-package org.permanent.permanent.ui.publicWorkspace
+package org.permanent.permanent.ui.public
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import kotlinx.android.synthetic.main.activity_main.*
-import org.permanent.permanent.R
-import org.permanent.permanent.databinding.FragmentEditMilestoneBinding
+import org.permanent.permanent.databinding.FragmentPublicGalleryBinding
 import org.permanent.permanent.ui.PermanentBaseFragment
 import org.permanent.permanent.viewmodels.PublicProfileViewModel
 
-class PublicProfileEditMilestonesFragment: PermanentBaseFragment(){
+class PublicGalleryFragment: PermanentBaseFragment() {
     private lateinit var viewModel: PublicProfileViewModel
-    private lateinit var binding: FragmentEditMilestoneBinding
+    private lateinit var binding: FragmentPublicGalleryBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentEditMilestoneBinding.inflate(inflater, container, false)
+        binding = FragmentPublicGalleryBinding.inflate(inflater, container, false)
         binding.executePendingBindings()
         binding.lifecycleOwner = this
-        activity?.toolbar?.menu?.findItem(R.id.settingsItem)?.isVisible = false
 
         return binding.root
     }
@@ -32,5 +29,4 @@ class PublicProfileEditMilestonesFragment: PermanentBaseFragment(){
 
     override fun disconnectViewModelEvents() {
     }
-
 }
