+++ app/src/main/java/org/permanent/permanent/viewmodels/AddEditOnlinePresenceViewModel.kt
@@ -14,7 +14,7 @@ import org.permanent.permanent.repositories.ProfileRepositoryImpl
 import org.permanent.permanent.ui.PREFS_NAME
 import org.permanent.permanent.ui.PreferencesHelper
 
-class AddEditSocialMediaViewModel(application: Application) : ObservableAndroidViewModel(application) {
+class AddEditOnlinePresenceViewModel(application: Application) : ObservableAndroidViewModel(application) {
     private val appContext = application.applicationContext
     private val prefsHelper = PreferencesHelper(
         application.getSharedPreferences(PREFS_NAME, Context.MODE_PRIVATE)
@@ -26,12 +26,14 @@ class AddEditSocialMediaViewModel(application: Application) : ObservableAndroidV
     private val socialMedia = MutableLiveData("")
     private var socialMediaProfileItem: ProfileItem? = null
     private var profileRepository: IProfileRepository = ProfileRepositoryImpl(application)
-    private var isOnEdit: Boolean? = false
+    private var isEdit: Boolean? = false
+    private var isAddEmail: Boolean? = false
 
-    fun displayProfileItem(profileItem: ProfileItem?, onEdit: Boolean?) {
+    fun displayProfileItem(profileItem: ProfileItem?, isEdit: Boolean?, isAddEmail: Boolean?) {
         socialMediaProfileItem = profileItem
         profileItem?.string1?.let { socialMedia.value = it }
-        onEdit?.let { isOnEdit = onEdit}
+        isEdit?.let { this.isEdit = isEdit}
+        isAddEmail?.let { this.isAddEmail = isAddEmail}
     }
 
     fun onSocialMediaTextChanged(text: Editable) {
@@ -52,7 +54,11 @@ class AddEditSocialMediaViewModel(application: Application) : ObservableAndroidV
             if (socialMedia.value?.trim()?.isNotEmpty() == true) {
                 val addItem = ProfileItem ()
                 addItem.archiveId = prefsHelper.getCurrentArchiveId()
-                addItem.fieldName = ProfileItemName.SOCIAL_MEDIA
+                if(isAddEmail ==  true){
+                    addItem.fieldName = ProfileItemName.EMAIL
+                } else {
+                    addItem.fieldName = ProfileItemName.SOCIAL_MEDIA
+                }
                 addItem.string1 = socialMedia.value?.trim()
                 addUpdateProfileItem(addItem)
             }
@@ -68,7 +74,7 @@ class AddEditSocialMediaViewModel(application: Application) : ObservableAndroidV
                 override fun onSuccess(profileItem: ProfileItem) {
                     isBusy.value = false
                     profileItemToUpdate.id = profileItem.id
-                    if(isOnEdit == true) {
+                    if(isEdit == true) {
                         showMessage.value = appContext.getString(R.string.edit_social_media_success)
                     }
                     else
@@ -86,6 +92,7 @@ class AddEditSocialMediaViewModel(application: Application) : ObservableAndroidV
     fun getShowMessage(): LiveData<String?> = showMessage
     fun getShowError(): LiveData<String?> = showError
     fun getOnBackRequest(): LiveData<Void> = onBackRequest
+    fun getIsOnAddEmail(): Boolean? = isAddEmail
 
     fun getSocialMedia(): LiveData<String> = socialMedia
 }
