+++ app/src/main/res/values/strings.xml
@@ -166,10 +166,10 @@
     <string name="edit_milestone_start_date">Start Date</string>
     <string name="edit_milestone_end_date">End Date</string>
     <string name="edit_milestone_description_hint">Describe this milestone</string>
-    <string name="edit_add_social_media_hint">www.website.com/user</string>
+    <string name="edit_add_social_media_hint">https://www.website.com/user</string>
+    <string name="edit_add_email_hint">user@example.com</string>
     <string name="edit_online_presence_label">Edit Online Presence</string>
-    <string name="add_email_button">Add email</string>
-    <string name="add_social_media_button">Add social media</string>
+    <string name="add_email_button">Add Email</string>
     <string name="add_social_media_label">Add Social Media</string>
     <string name="add_social_media_success">Social Media added successfully</string>
     <string name="edit_social_media_success">Social Media edited successfully</string>
